# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: BabyAI API
  version: 0.1.0
  description: >
    BabyAI MVP API. 육아 기록, AI 조회, 리포트, 구독/사진 공유, Siri/Bixby 연동을 포함한다.
servers:
  - url: http://localhost:8000
    description: local
...
paths:
...
  /api/v1/photos/upload-url:
    post:
      tags: [photos]
      summary: Create presigned upload URL
      operationId: createPhotoUploadUrl
      parameters:
        - in: query
          name: album_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Presigned URL created
          content:
            application/json:
              schema:
                type: object
                properties:
                  album_id: { type: string }
                  upload_url: { type: string, format: uri }
                  object_key: { type: string }
  /api/v1/photos/complete:
    post:
      tags: [photos]
      summary: Confirm photo upload and generate variants metadata
      operationId: completePhotoUpload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoUploadCompleteRequest'
      responses:
        '200':
          description: Upload completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: uploaded }
                  photo_id: { type: string }
                  album_id: { type: string }
                  downloadable: { type: boolean }
                  variants:
                    type: object
                    additionalProperties: { type: string }
...
  /api/v1/subscription/me:
    get:
      tags: [subscription]
      summary: Get household subscription
      operationId: getMySubscription
      parameters:
        - in: query
          name: household_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Subscription snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  household_id: { type: string }
                  plan: { $ref: '#/components/schemas/SubscriptionPlan' }
                  status: { type: string, example: active }
  /api/v1/subscription/checkout:
    post:
      tags: [subscription]
      summary: Start checkout
      operationId: checkoutSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: pending_payment }
                  plan: { $ref: '#/components/schemas/SubscriptionPlan' }
                  household_id: { type: string }
...
components:
  schemas:
...
    SubscriptionPlan:
      type: string
      enum: [AI_ONLY, AI_PHOTO]
...
    PhotoUploadCompleteRequest:
      type: object
      required: [album_id, object_key]
      properties:
        album_id: { type: string }
        object_key: { type: string }
        downloadable: { type: boolean, default: false }
...
    CheckoutRequest:
      type: object
      required: [household_id, plan]
      properties:
        household_id: { type: string }
        plan: { $ref: '#/components/schemas/SubscriptionPlan' }
...